moved to c++ implementation

+++ apps/ext/sunsky/ArHosekSkyModel.c
@@ -523,7 +523,7 @@ double arhosekskymodel_radiance(
         double                  wavelength
         )
 {
-    int low_wl = (wavelength - 320.0 ) / 40.0;
+    int low_wl = (int)((wavelength - 320.0f ) / 40.0f);
 
     if ( low_wl < 0 || low_wl >= 11 )
         return 0.0f;
